Move PATH related explanations on Windows Install Instructions to the top (#145)
* doc: Move explanation at the top 

How to extend the PATH variable may be explained at first occurance (line 51). In documentation it is explanined at lines 164 and 165. However it is used before (lines 51, 148, 155)

* fix: to your PATH

change "to the PATH" to "to your PATH" as it may change from user to user

* rephrase instruction

rephrase instuction at line 47 to be consistent with following lines.
add to "your" path since path change from user to user

* remove paranthesis

remove paranthesis as it includes more than one sentence. change "then" word at the end to "finally".

* one sentence per line

Co-Authored-By: mehmety888 <mehmety888@gmail.com>

* remove whitespaces

@hidmic no problem, please check again.