Use CMAKE_CXX_STANDARD in the migration guide (#283)
* Use CMAKE_CXX_STANDARD in the migration guide

CMAKE_CXX_STANDARD is the preferred way to state the required C++ standard as it works across all compilers supported by CMake.

* Require C++14 instead of C++11 in the migration guide

* Updated more occurrences of C++11 in the migration guide

In the case of the boost transition, it doesn't really matter which standard this is, only that they have been integrated and should be replaced. We can therefore remove the year.