Updates to the Fedora development setup. (#289)
Updates to support the build-from-source instructions.  Tested by
following the directions on the wiki in a fedora:30 docker container.

Update the package list to use dnf to install python3-colcon-common
extensions, which is available in Fedora 29 and newer:
https://bodhi.fedoraproject.org/updates/FEDORA-2019-8c1b78e3f3

wget and rosdep are required by the build from source instructions, and
not installed by default.

gcc-c++, patch, and make are required to build C++ packages, and may not
be installed on a fresh system. redhat-rpm-config provides hardening
flags that gcc-c++ needs.

liblsan is needed to build rosbag2 (the sanitizer isn't installed by
default when gcc-c++ is installed).  It should probably be disabled by
default, and/or added as a system rosdep.

Remove pip command - all relevant dependencies can be installed with the
system package manager.

Also, a couple of cosmetic updates:
- Put each package on a new line to match the Linux Development Setup
page formatting
- Alphabatize package list
- Re-word first paragraph to reflect that the setup is now a single
step.

Signed-off-by: Rich Mattes <richmattes@gmail.com>