Add pydot to the macOS from source instructions. (#669)
* Add pydot to the macOS from source instructions.

This has been needed all the way back to Dashing, and *is*
included in the binary instructions.  It was just missing
from the source instructions.

The RQt dependencies were previously included only in the binary
installation instructions. Except on Eloquent where there was an
incomplete set (graphviz and pygraphviz were missing).

Signed-off-by: Chris Lalancette <clalancette@openrobotics.org>
Signed-off-by: Steven! Ragnarök <steven@nuclearsandwich.com>
Co-authored-by: Steven! Ragnarök <steven@nuclearsandwich.com>